Feat: connext

+++ app/src/main/java/com/sas/companymanagement/ui/artist/db/ArtistRepository.kt
@@ -60,7 +60,7 @@ class ArtistRepository(application: Application) {
     }
 
 
-    fun findArtistById(id: Int): LiveData<Artist> {
+    fun findArtistById(id: Long): LiveData<Artist> {
         return artistDao.findArtistById(id)
     }
     fun updateUser(artist: Artist) {
